feat(products): Alterado tipo na coluna price na migration de Produtos

+++ api/database/migrations/2024_02_28_212556_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price');
+            $table->decimal('price');
             $table->text('description');
             $table->timestamps();
             $table->softDeletes();
